pipeline/orchestration/dag.py
"""
Apache Airflow DAG for Smart Meter MDMS Pipeline
Orchestrates: ingestion -> cleaning -> features -> validation -> storage -> SQL -> reports
"""

from datetime import datetime, timedelta
from pathlib import Path
import logging

logger = logging.getLogger(__name__)

# This DAG definition is for reference and local execution
# In production, place this file in AIRFLOW_HOME/dags/


class SmartMeterDAG:
    """
    DAG definition for Smart Meter MDMS Pipeline.
    Orchestrates complete data flow from raw ingestion to final reports.
    """
    
    def __init__(self):
        """Initialize DAG configuration."""
        self.dag_id = 'smart_meter_mdms_pipeline'
        self.default_args = {
            'owner': 'smart_meter_team',
            'depends_on_past': False,
            'start_date': datetime(2026, 1, 1),
            'email_on_failure': False,
            'email_on_retry': False,
            'retries': 1,
            'retry_delay': timedelta(minutes=5),
        }
        
        self.schedule_interval = '@daily'
        self.description = 'Smart Meter MDMS complete pipeline orchestration'
        self.catchup = False
        
        logger.info(f"Initialized DAG: {self.dag_id}")
    
    def get_task_dependencies(self) -> dict:
        """
        Define task dependencies for the pipeline.
        
        Returns:
            Dictionary describing task flow
        """
        return {
            'ingestion': {
                'id': 'ingest_raw_data',
                'description': 'Load raw smart meter data from CSV',
                'depends_on': None,
                'function': 'src.ingestion.RawDataIngestion.ingest_and_prepare'
            },
            'cleaning': {
                'id': 'clean_data',
                'description': 'Clean data: dedup, impute, fix invalid values',
                'depends_on': ['ingestion'],
                'function': 'src.cleaning.DataCleaner.clean_pipeline'
            },
            'feature_engineering': {
                'id': 'engineer_features',
                'description': 'Add temporal, electrical, aggregation features',
                'depends_on': ['cleaning'],
                'function': 'src.feature_engineering.FeatureEngineer.feature_engineering_pipeline'
            },
            'validation': {
                'id': 'validate_data',
                'description': 'Validate completeness, uniqueness, validity, consistency',
                'depends_on': ['feature_engineering'],
                'function': 'src.validation.DataValidator.validation_pipeline'
            },
            'storage': {
                'id': 'store_processed',
                'description': 'Save processed data to CSV, Parquet, CuratedStorage',
                'depends_on': ['validation'],
                'function': 'src.storage.DataStorage.store_all'
            },
            'sql_execution': {
                'id': 'execute_sql',
                'description': 'Create database, insert data, run analytics queries',
                'depends_on': ['storage'],
                'function': 'src.sql_runner.SQLRunner.execute_sql_file'
            },
            'spark_processing': {
                'id': 'spark_process',
                'description': 'Transform data using Spark / Pandas fallback',
                'depends_on': ['storage'],
                'function': 'pipeline.processing.spark_processor.SparkProcessor.process'
            },
            'hdfs_archival': {
                'id': 'archive_to_hdfs',
                'description': 'Archive processed data to HDFS simulation',
                'depends_on': ['storage'],
                'function': 'pipeline.hadoop.hdfs_manager.HDFSManager.put'
            },
            'eda_analysis': {
                'id': 'generate_eda',
                'description': 'Generate EDA plots and statistics',
                'depends_on': ['storage'],
                'function': 'src.eda_analysis.SmartMeterEDA.generate_all_plots'
            },
            'reporting': {
                'id': 'create_reports',
                'description': 'Generate final analytics reports and summary',
                'depends_on': ['eda_analysis', 'sql_execution'],
                'function': 'src.analytics_engine.AnalyticsEngine.generate_reports'
            }
        }
    
    def get_task_order(self) -> list:
        """
        Get logical execution order of tasks.
        
        Returns:
            List of task IDs in execution order
        """
        return [
            'ingestion',
            'cleaning',
            'feature_engineering',
            'validation',
            'storage',
            'spark_processing',
            'hdfs_archival',
            'sql_execution',
            'eda_analysis',
            'reporting'
        ]


# For Airflow to pick up: try to create actual DAG if airflow is available
try:
    from airflow import DAG
    from airflow.operators.python import PythonOperator
    from airflow.operators.bash import BashOperator
    from airflow.utils.decorators import dag
    
    dag_config = SmartMeterDAG()
    
    @dag(
        dag_id=dag_config.dag_id,
        default_args=dag_config.default_args,
        schedule_interval=dag_config.schedule_interval,
        description=dag_config.description,
        catchup=dag_config.catchup,
    )
    def smart_meter_mdms_dag():
        """
        Smart Meter MDMS complete pipeline DAG.
        """
        
        # Task definitions would go here
        # This is a reference implementation
        
        pass
    
    # Create DAG
    smart_meter_mdms_dag = smart_meter_mdms_dag()
    logger.info("Airflow DAG created successfully")

except ImportError:
    logger.info("Airflow not available - DAG is defined for reference only")


def main():
    """Example usage."""
    dag_config = SmartMeterDAG()
    logger.info(f"DAG Config: {dag_config.dag_id}")
    logger.info(f"Tasks: {list(dag_config.get_task_dependencies().keys())}")


if __name__ == "__main__":
    main()
