requirements.txt
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1
scipy==1.11.1
scikit-learn==1.3.0
SQLAlchemy==2.0.19
python-dateutil==2.8.2
pyspark==3.5.0
confluent-kafka==2.3.0
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.1.1
psycopg2-binary==2.9.9
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
jupyter==1.0.0
ipython==8.14.0
